Switch selenium for capybara-webkit for cucumber
* Speed up tests
* Fix broken cucumber step definition
* Put the config for block_unknown_urls in an actual config

The use of `page.should have_content("")` might seem a bit weird. It is used as
a synchronization point, to make sure that the previous request has finished
before making any further assertions in that scenario.  This is necessary since
the necessary changes will not have been persisted until that previous request
is done, and `capybara-webkit` apparently is a bit too fast sometimes.