Reconfigure how to run slow specs
Still default to not running them, but move the configuration into `spec_helper`
instead of having it in `.rspec`.

This way, we can run them conditionally if either `ENV["CI"]` or
`ENV["RUN_ALL"]` is set. Travis CI sets the `CI` environment variable by
default, so that would make all the specs run on travis.