Add note about pg_trgm set_limit query
The set_limit(0.2) query is run to change the pg_trgm threshold in order to
match some partial (short) words that the default didn't catch.

This threshold is a per-db-connection thing, and that is why it's done here,
instead of somewhere else. The textacular crowd don't want to add it as a
configuration option since that would probably only work in Rails land, and they
want to be compatible with other parts of the ruby community as well.